[IWA] Define IWA build resources
Since the source code will be compiled into PWA and IWA, we will
define two separate folders for them.

The PWA distribution code will be located in `/dist/pwa`
The IWA distribution web bundle will be located in `/dist/iwa`

There are certain files that are IWA specifc. For now, that include
the IWA manifest and the png icon for the app. These are placed in a
separate folder `iwa/public`.

BUG: b/295127885