[UI] Implement TelemetryDashboard Component
An array is declared in file telemetry-dashboard.component.ts to specify
which telemetry types should be displayed in the dashboard. And the
program loops through those types and uses telemetry service to request
and receive the desired data.

And the TelemetryCardComponent is used in the template for showing
telemetry base cards.

BUG: b/295126796