[UI] Change data refresh interval format
Since there will be refresh intervals for fetching telemetry info,
diagnostics routine update, and events update, the refreshIntervals is
changed from a map that can be accessed using telemetry info types to an
object that has a key 'telemetry' to a value with type
Map<string, number> that uses telemetry info type as key and refresh
interval as value.

And a variable defaultTelemetryRefreshInterval is set as a default value
of refresh interval.

BUG: b/295126796