[UI] Create TelemetryCard Component
The TelemetryCardComponent is created, and it takes one input: type,
which specifies which type of telemetry info it is fetching.

As for the content of the card, when it is fetching the data, a loading
progress spinner will be shown, and when there is an error, a yellow
error message will be displayed, and when the data is successfully
aquired, it renders TelemetryCardContentComponent and sends the data as
input.

BUG: b/295126796