[UI] Implement DiagnosticsDashboard Component
An array is declared in the ts file to set which diagnostics routines
should be displayed on the diagnostics dashboard. And it calls the
service function getAvailableRoutines() in advance to check whether the
routines are available. Once it receives the response, the program
filters the routines and renders the diagnostics card for each available
routine.

And if there is an error occurs during the service function call, the
error message will be shown on the dashboard.

BUG: b/295126759