[UI] Create DiagnosticsCard Component
The DiagnosticsCardComponent is created and it takes one input: routine,
which specifies the diagnostics routine type the card represents for.

And the component contains a header bar, a progress bar, and a message
bar. The header bar shows the routine name, the its parameters for
running the routine, and three buttons that starts, stops, or resumes
the routine when clicked.

When a routine starts, the program call the service for getting routine
status according to its refresh interval, and when it receives the
status, it updates the progress bar and shows the progress percentage or
status message on the card. And once the routine reaches the terminal
states, it presents the corresponding status (e.g. Passed, Failed, etc.)

BUG: b/295126759