[RMA] Add Logging service
Add a logging service which will allow logs to be stored and
processed.

This is needed for RMA mode where warning/error logs will be shown in
the dedicated log-viewer area.

The service is provided in top level module, allowing the same logs to
be shared and accessed across the codebase.

BUG: b/326506274