[RMA] Add RMA UI sidebar
Add a sidebar in RMA mode. This sidebar should reflect the tests that
are defined in the testlist.

The user should be abel to select or deselect tests by clicking the
checkbox.

Clicking on each test should redirect to the relevant test page.

Below shows the UI for the RMA sidebar:
https://screenshot.googleplex.com/6z7wxEcwgx7ucnk